ATOM-15518 Change Multilayer PBR To Use Lerp Base Blending
Updated test cases to accomodate changes to the engine.

Changed to lerp-based blending, with an implicit base layer.
Updated DefaultBlendMask_layers.png to better suit the new layering model.
Added a new test object, which is a plane that has painted vertices. Changed test cases to use this instead of the "com_envi_me_wall-2x2_testcorner" test object which was painted according to the old layer blending strategy.