Add Soak Test To Validate reload notifications of Shader related assets
Added new Example called "RPI/ShaderReloadSoakTest".
As an example it provides an ImGUI UI to change the
shader used for the FullscreenTrianglePass which renders
a solid Red, Green or Blue color.

A button to check the expected color is vaialble.

A lau script is available: ShaderReloadSoakTest.bv.lua, which
mimics pressing the buttons and testing for the expected color
when the shader asset changes.

Because this is a new Soak Test it should NOT be added
to any of the automated test suites.

Signed-off-by: galibzon <66021303+galibzon@users.noreply.github.com>