Test shader definitions field in shader reload soak test
This commit adds a new test shader file FullscreenMacroWhite.shader.txt
which adds a preprocessor definition "WHITE" to the shader. The red,
blue, and green shader source files all conditionally emit 0xFFFFFFFF if
this macro is defined, and the shader reload test checks for a white pixel
if this shader is used.

Signed-off-by: Jeremy Ong <jcong@amazon.com>