Adding readback sample
This commit introduce the readback sample. It is designed to render a
pattern into a texture that is then readback to host memory and
reuploaded to the GPU to be rendered.
This is only the shell of the proper test that will be updated as soon
as the changes supporting subresource readback are added.

Signed-off-by: Bindless-Chicken <1039134+Bindless-Chicken@users.noreply.github.com>