ATOM-17550 RPI/BakedShaderVariant: Sample crashes when exiting the application
The post effect entity was deleted twice if exist the app when the sample is running.
Remove the entity since it's not necessary.

The idea pattern to delete an entity created from the component, is to save the entity id and remove the entity by id. We should change all ASV code using this pattern.
This can simplify the code which we shouldn't need to connect to EntityBus and listen to OnEntityDestruction() event anymore.

Signed-off-by: Qing Tao <55564570+VickyAtAZ@users.noreply.github.com>