Do not create the default render pipeline when using xr systems. (#9)
Instead of disabling the default render pipeline when it's not needed (when running on a XR system), this change avoids its creation at all, saving resources.

Approach:
On host platforms (windows, linux, mac) the render pipeline is always created (with or without xr system), and when there is an xr system the host render pipeline can be enabled/disabled using `r_EnableHostRenderPipelineOnXR` cvar.

On non-host platforms the render pipeline will not be created when there is an xr system.

Signed-off-by: moraaar <moraaar@amazon.com>
Signed-off-by: amzn-phist <52085794+amzn-phist@users.noreply.github.com>